+++ iso20022-core/src/test/java/com/prowidesoftware/swift/model/mx/adapters/MxWriteWithAdaptersTest.java
@@ -44,9 +44,9 @@ public class MxWriteWithAdaptersTest {
         final String xml = mx1.message();
         // System.out.println(xml);
         assertTrue(xml.contains(
-                "<Doc:CreDtTm>2021-10-19T12:13:14" + OffsetDateTime.now().getOffset() + "</Doc:CreDtTm>"));
-        assertTrue(xml.contains("<Doc:IntrBkSttlmDt>2021-10-19</Doc:IntrBkSttlmDt>"));
-        assertTrue(xml.contains("<Doc:CLSTm>12:13:14" + OffsetDateTime.now().getOffset() + "</Doc:CLSTm>"));
+                "<pacs:CreDtTm>2021-10-19T12:13:14" + OffsetDateTime.now().getOffset() + "</pacs:CreDtTm>"));
+        assertTrue(xml.contains("<pacs:IntrBkSttlmDt>2021-10-19</pacs:IntrBkSttlmDt>"));
+        assertTrue(xml.contains("<pacs:CLSTm>12:13:14" + OffsetDateTime.now().getOffset() + "</pacs:CLSTm>"));
 
         final MxPacs00800102 mx2 = MxPacs00800102.parse(xml);
         // System.out.println(mx2.message());
@@ -95,9 +95,9 @@ public class MxWriteWithAdaptersTest {
 
         final String xml = mx1.message();
         // System.out.println(xml);
-        assertTrue(xml.contains("<Doc:CreDtTm>2021-10-19T12:13:14+00:00</Doc:CreDtTm>"));
-        assertTrue(xml.contains("<Doc:IntrBkSttlmDt>2021-10-19</Doc:IntrBkSttlmDt>"));
-        assertTrue(xml.contains("<Doc:CLSTm>12:13:14+00:00</Doc:CLSTm>"));
+        assertTrue(xml.contains("<pacs:CreDtTm>2021-10-19T12:13:14+00:00</pacs:CreDtTm>"));
+        assertTrue(xml.contains("<pacs:IntrBkSttlmDt>2021-10-19</pacs:IntrBkSttlmDt>"));
+        assertTrue(xml.contains("<pacs:CLSTm>12:13:14+00:00</pacs:CLSTm>"));
 
         final MxPacs00800102 mx2 = MxPacs00800102.parse(xml);
         // System.out.println(mx2.message());
@@ -147,9 +147,10 @@ public class MxWriteWithAdaptersTest {
         final String xml = mx1.message();
         // System.out.println(xml);
         assertTrue(xml.contains(
-                "<Doc:CreDtTm>2021-10-19T12:13:14.123" + OffsetDateTime.now().getOffset() + "</Doc:CreDtTm>"));
-        assertTrue(xml.contains("<Doc:IntrBkSttlmDt>2021-10-19</Doc:IntrBkSttlmDt>"));
-        assertTrue(xml.contains("<Doc:CLSTm>12:13:14.123" + OffsetDateTime.now().getOffset() + "</Doc:CLSTm>"));
+                "<pacs:CreDtTm>2021-10-19T12:13:14.123" + OffsetDateTime.now().getOffset() + "</pacs:CreDtTm>"));
+        assertTrue(xml.contains("<pacs:IntrBkSttlmDt>2021-10-19</pacs:IntrBkSttlmDt>"));
+        assertTrue(
+                xml.contains("<pacs:CLSTm>12:13:14.123" + OffsetDateTime.now().getOffset() + "</pacs:CLSTm>"));
 
         final MxPacs00800102 mx2 = MxPacs00800102.parse(xml);
         // System.out.println(mx2.message());
@@ -202,8 +203,8 @@ public class MxWriteWithAdaptersTest {
 
         final String xml = mx1.message(conf);
         // System.out.println(xml);
-        assertTrue(xml.contains("<Doc:CreDtTm>21-10-19 12:13</Doc:CreDtTm>"));
-        assertTrue(xml.contains("<Doc:IntrBkSttlmDt>21-10-19</Doc:IntrBkSttlmDt>"));
+        assertTrue(xml.contains("<pacs:CreDtTm>21-10-19 12:13</pacs:CreDtTm>"));
+        assertTrue(xml.contains("<pacs:IntrBkSttlmDt>21-10-19</pacs:IntrBkSttlmDt>"));
 
         final MxPacs00800102 mx2 = MxPacs00800102.parse(xml, new MxReadConfiguration(conf));
         // System.out.println(mx2.message());
@@ -251,8 +252,8 @@ public class MxWriteWithAdaptersTest {
 
         final String xml = mx1.message(conf);
         // System.out.println(xml);
-        assertTrue(xml.contains("<Doc:CreDtTm>foobar</Doc:CreDtTm>"));
-        assertTrue(xml.contains("<Doc:IntrBkSttlmDt>foobar</Doc:IntrBkSttlmDt>"));
+        assertTrue(xml.contains("<pacs:CreDtTm>foobar</pacs:CreDtTm>"));
+        assertTrue(xml.contains("<pacs:IntrBkSttlmDt>foobar</pacs:IntrBkSttlmDt>"));
 
         final MxPacs00800102 mx2 = MxPacs00800102.parse(xml, new MxReadConfiguration(conf));
         // System.out.println(mx2.message());