+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   * SWIFT Standard release update 2024 (live 16 November 2024)
   * Yearly revision of deprecation phase (see https://dev.prowidesoftware.com/SRU2024/getting-started/deprecation/)
   * Added message categories for File Management (cafm), Fraud Reporting and Disposition (cafm), Network Management (canm) and Settlement Reporting (casr)
+  * Add support in the MxWriteConfiguration to use standard envelopes for SWIFT and ISO 20022
+  * Changed the default Document prefix to the message category; "camt", "pacs", etc...
 
 #### 9.4.4 - January 2024
   * Enhanced the identifier extraction of the MxSwiftMessage to use the AppHdr when the Document namespace is missing